Move transformers and related classes to their own folder.
