+++ SolusEngine.cs
@@ -35,6 +35,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using MetaphysicsIndustries.Solus.Expressions;
 using MetaphysicsIndustries.Solus.Functions;
+using MetaphysicsIndustries.Solus.Transformers;
 
 namespace MetaphysicsIndustries.Solus
 {
