+++ Transformers/CloneTransformer.cs
@@ -20,13 +20,20 @@
  *
  */
 
-using System;
-using System.Collections.Generic;
-using System.Text;
+using MetaphysicsIndustries.Solus.Expressions;
 
-namespace MetaphysicsIndustries.Solus
+namespace MetaphysicsIndustries.Solus.Transformers
 {
-    public abstract class TransformArgs 
+    public class CloneTransformer : ExpressionTransformer
     {
+        public override bool CanTransform(Expression expr)
+        {
+            return (expr != null);
+        }
+
+        public override Expression Transform(Expression expr)
+        {
+            return expr.Clone();
+        }
     }
 }
