+++ Transformers/RatioOfPolynomialsTransformer.cs
@@ -20,13 +20,11 @@
  *
  */
 
-using System;
 using System.Collections.Generic;
-using System.Text;
 using MetaphysicsIndustries.Solus.Expressions;
 using MetaphysicsIndustries.Solus.Functions;
 
-namespace MetaphysicsIndustries.Solus
+namespace MetaphysicsIndustries.Solus.Transformers
 {
     public class RatioOfPolynomialsTransformer : PolynomialTransformer
     {
